Initial Commit. Severely lacking
Here is an initial framework for Sys-update. Only Debian based systems
are somewhat worked through. It will update the repositires and the
groundwork is laid for the rest but there is still much to do. Once
debian gets fully working, it will be rather simple to get arch and rhel
working after that. I may even put in support for solus because it will
be that easy. Still to do: Make a screen showing available updates (they
available updates are alreadyfound, just need a new screen to put them
on), an update screen, and a sys info screen. All of these can be the
same thing for all types of distros, just use different processes and
information depending on what they are.